git-svn-id: https://svn.r-project.org/R-packages/trunk/Mac-GUI@5707 694ef91d-65df-0310-b7bb-92e67a308ead

+++ RConsoleTextStorage.m
@@ -56,7 +56,7 @@
 	return [cont string];
 }
 
-- (NSDictionary *) attributesAtIndex:(unsigned)index effectiveRange:(NSRangePointer)aRange
+- (NSDictionary *) attributesAtIndex:(NSUInteger)index effectiveRange:(NSRangePointer)aRange
 {
 	return [cont attributesAtIndex:index effectiveRange:aRange];
 }
